Move integration tests into new test module
