+++ zrml/prediction-markets/src/lib.rs
@@ -26,7 +26,6 @@ extern crate alloc;
 mod benchmarks;
 pub mod migrations;
 pub mod mock;
-mod old_tests;
 pub mod orml_asset_registry;
 mod tests;
 pub mod weights;
